platformer_tutorial_2025/pyproject.toml
[project]
name = "platformer-tutorial-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "Cooper Park"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce (>=2.5.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
